[fe notes] setup
