+++ .gitignore
@@ -102,3 +102,25 @@ dist
 
 # TernJS port file
 .tern-port
+
+## IDEs / editors
+.idea/
+*.iml
+.project
+.classpath
+.c9/
+*.launch
+.settings/
+*.sublime-workspace
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+.history/*
+# emacs
+*~
+\#*\#
+#.\#*
+
+
