+++ frontend/src/components/App.css
@@ -1,0 +1,11 @@
+.App {
+  text-align: center;
+
+}
+
+.App h4 {
+  margin: 2em;
+  padding: 4em;
+  border-style: solid;
+  border-width: medium;
+}
