+++ frontend/src/components/App.test.js
@@ -1,0 +1,8 @@
+import { render, screen } from '@testing-library/react';
+import App from './App';
+
+test('renders nle text', () => {
+  render(<App />);
+  const nleElement = screen.getByText(/nle/i);
+  expect(nleElement).toBeInTheDocument();
+});
