+++ frontend/src/components/notes/NotesSteward.js
@@ -1,0 +1,9 @@
+function NotesSteward () {
+    return (
+        <div>
+            notes...
+        </div>
+    );
+}
+
+export default NotesSteward;
